refactor: step5 Winner 클래스 분리
