+++ src/main/java/step5/domain/GameRound.java
@@ -7,12 +7,10 @@ import step5.exception.RacingGameException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class GameRound {
 
     private List<ResultOfCar> gameRoundResult;
-    private List<String> winner;
 
     void play(List<Car> carList, CarRacingGameStrategy carRacingGameStrategy) {
         gameRoundResult = new ArrayList<>();
@@ -26,30 +24,9 @@ public class GameRound {
 
         }
 
-        decideWinner();
-
     }
 
-    private void decideWinner() {
-        int max = gameRoundResult.stream()
-                .mapToInt(ResultOfCar::getPosition)
-                .max()
-                .orElseThrow(() -> new RacingGameException("winner가 없습니다"));
 
-        winner = gameRoundResult.stream()
-                .filter(r -> (r.getPosition() == max))
-                .map(ResultOfCar::getCarName)
-                .collect(Collectors.toList());
-
-    }
-
-    public List<String> getWinners() {
-        if (winner == null) {
-            throw new RacingGameException("아직 게임을 진행하지 않았습니다");
-        }
-        return Collections.unmodifiableList(winner);
-
-    }
 
     public List<ResultOfCar> getGameRoundResult() {
         if (gameRoundResult == null) {
@@ -57,6 +34,5 @@ public class GameRound {
         }
         return Collections.unmodifiableList(gameRoundResult);
     }
-
 
 }
